Catch possible exception in https download test.
It happened, due to the binary SSL protocol, that the connection
is closed from the download client before the dummy message could be
sent. This results in a Errno::ECONNRESET in the server thread.
The thread is therefore terminated silently, so that the download
client can not establish the next connection attempt.
This freezed the test.